backend/requirements.txt
# HistoryBuff Backend Dependencies
#
# Core framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.25
alembic>=1.13.0
pgvector>=0.2.4

# Task queue
celery>=5.3.6
redis>=5.0.1
flower>=2.0.1

# AI/ML
anthropic>=0.18.0
openai>=1.10.0
sentence-transformers>=2.3.0
torch>=2.1.0  # CPU only, for sentence-transformers

# Extraction
pytesseract>=0.3.10
Pillow>=10.2.0
pypdf2>=3.0.1
beautifulsoup4>=4.12.0
lxml>=5.1.0

# HTTP client
httpx>=0.26.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
python-multipart>=0.0.6
python-magic>=0.4.27
tenacity>=8.2.0  # Retry logic

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
ruff>=0.1.0
mypy>=1.8.0
